requirements.txt
flask
langchain
langchain-community
langchain-google-genai
langchain-pinecone
pinecone-client
sentence-transformers
python-docx
unstructured[docx]
pypdf
